Allow user to specify fallback and fix #6
If no magic comment was found, TeXMagicBuildConfig is currently not set
(see #6). This commit adds logic in the branch if no magic comments were
not found.

Furthermore, it allows the user to specify a default engine that should
be used if no comment was found or the request tex program was not
found.